bot/requirements.txt
vk-api
python-dotenv
watchdog
apscheduler
requests
